Fixed an issue related to improper flag calculation for crafted packets. Flags should always be correct now. miniPackets also do not count it's size bytes in it's total size calculation. Removed the requirement of specifying the flaglength for buildPacket and buildMiniPacket, all packet flagLengths should be declared explicitly within AlefFlagLengthLookup. Updated more client login packets, and as a result we can now get to the character list, albeit it is currently unpopulated. The next step is to populate it with some dummy data. Some cosmetic changes were also done to the AlefStartupEncryption class, AlefPacketInterface, and more organization was implemented in the project files.
Note: Due to the class hierarchy, the AlefLoginPacketHandler (and by extension the AlefWorldPacketHandler) does not destruct until the server is given the close command, and as a result we are unable to specify packets within the packet handler classes themselves, as building new packets creates new memory and they can never destruct giving us memory leaks. From this, we are forced to continue with the current methodology of crafting packets within the specific handler functions and deleting them at the end. This will remain the way ALEmu does things until a better solution comes along, either in the direction altering the class hierarchy, or by handling packet memory in a different way.