- AlefPacketInterface updates, this should finalize it. We can read in and process miniPackets now, though further processing is required beyond the initial read in. - Foundations of the "local data" system. Essentially any data unique to a client connection object will be stored within it's own connection context and we will work through that data as needed. - Added support for EntityTemplates and EntityData, which are the charactertemplatepublic.ini and characterdatatable.txt files respectively. - Added dependency copying for the new MySQL DLLs - Accounts now use SHA2 for their password hashing, there is a new account added test2:test that showcases this. - HashKey is now automatically generated for the user, for some reason only HashKeys of 8 bytes work, tho the protocol is supposed to use HashKeys of length 16. I don't think it matters much, but it's still something to note. - Added capability to send login results. - Various other improvements/cleanups.
Note: Part of the reason things have been progressing slowly is I can't figure out an elegant way of handling character packets and serializing them to cleanly give back to the packet handler classes to be sent. There currently is the beginnings of a solution I am okay with, but I'm still stuck in a few places. I'd like to at least have character creation done by the end of the month, but we will see...